Amend subcommand instantiation so option parsing is called only once.
This is *a* possible for for the scenario where we use the option
parsing to return an object that we don't want re-parse such as a config
file.

For example (in blah.rb):
require 'clamp'
require 'zlib'

class Moo < Clamp::Command
  option '--blob', 'BLOB', 'some blob' do |option|
    zstream = Zlib::Inflate.new
    buf = zstream.inflate(option)
    zstream.finish
    zstream.close
    buf
  end

  subcommand 'show-blob', 'show the blob' do
    def execute
      puts "whatevs"
    end
  end
end

When we run `blah.rb --blob asdf moo` we get an error as clamp will call
the block attached to the --blob option twice.

The fix in this case is to set the option on the subcommand directly
using instance_variable_set() - not the cleanest way :S